video: tegra: host: Don't depend on 32-bit ulong
Avoid BITS_PER_LONG and macros BIT_WORD and BIT_MASK that depend on
it. Also avoid ulong when a 32-bit value is expected.

Bug 1377229

Reviewed-on: http://git-master/r/298698
Reviewed-on: http://git-master/r/298672

+++ drivers/video/tegra/host/gk20a/fifo_gk20a.c
@@ -943,7 +943,7 @@ static bool gk20a_fifo_set_ctx_mmu_error(struct gk20a *g,
 static bool gk20a_fifo_handle_mmu_fault(struct gk20a *g)
 {
 	bool fake_fault;
-	unsigned long fault_id;
+	u32 fault_id;
 	u32 engine_mmu_id;
 	int i;
 	bool verbose = true;
